tests: add OpenZeppelin ERC1155 tests

+++ contracts/openzeppelin/access/AccessControl.sol
@@ -98,7 +98,7 @@ abstract contract AccessControl is Context, IAccessControl, ERC165 {
                 string(
                     abi.encodePacked(
                         "AccessControl: account ",
-                        Strings.toHexString(uint160(account), 20),
+                        Strings.toHexString(uint168(account), 20),
                         " is missing role ",
                         Strings.toHexString(uint256(role), 32)
                     )
