+++ contracts/openzeppelin/utils/structs/EnumerableSet.sol
@@ -222,7 +222,7 @@ library EnumerableSet {
      * already present.
      */
     function add(AddressSet storage set, address value) internal returns (bool) {
-        return _add(set._inner, bytes32(uint256(uint160(value))));
+        return _add(set._inner, bytes32(uint256(uint168(value))));
     }
 
     /**
@@ -232,14 +232,14 @@ library EnumerableSet {
      * present.
      */
     function remove(AddressSet storage set, address value) internal returns (bool) {
-        return _remove(set._inner, bytes32(uint256(uint160(value))));
+        return _remove(set._inner, bytes32(uint256(uint168(value))));
     }
 
     /**
      * @dev Returns true if the value is in the set. O(1).
      */
     function contains(AddressSet storage set, address value) internal view returns (bool) {
-        return _contains(set._inner, bytes32(uint256(uint160(value))));
+        return _contains(set._inner, bytes32(uint256(uint168(value))));
     }
 
     /**
@@ -260,7 +260,7 @@ library EnumerableSet {
      * - `index` must be strictly less than {length}.
      */
     function at(AddressSet storage set, uint256 index) internal view returns (address) {
-        return address(uint160(uint256(_at(set._inner, index))));
+        return address(uint168(uint256(_at(set._inner, index))));
     }
 
     /**
